backend/requirements/requirement.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
pillow
python-dotenv
loguru
djangorestframework-simplejwt
django-allauth 
dj-rest-auth
django-filter
